requirements.txt
click==8.1.8
# Keep crewai at version 0.130.0
# Other package versions are flexible and can be changed
# Only change crewai version if there are critical dependency conflicts that cannot be resolved by other means
crewai==0.130.0              
crewai-tools==0.47.1         
fastapi==0.110.3
uvicorn==0.29.0              
openai==1.93.0               
litellm==1.72.0              
langchain-core>=0.2.0
langchain-community>=0.2.0
pypdf                        
celery[redis]==5.3.6         
redis==5.0.4                 
flower==2.0.1                
python-dotenv==1.0.1         
pydantic>=2.5.0     
python-multipart         